requirements.txt
ml-dtypes==0.2.0
jax==0.4.11
jax-metal==0.0.4
flax==0.7.4
optax==0.1.7
orbax-checkpoint==0.4.1
matplotlib==3.8.0
